team_deploy/app/requirements.txt
Flask==1.1.2
flask-restplus==0.13.0
Werkzeug==0.16.1
pandas==1.2.4
scikit-learn==0.22.2
numpy~=1.19.2
tensorflow==2.5.0
scikit-image>=0.18.0
datetime
opencv-python-headless
Pillow==8.2.0